+++ module-baseinfo/src/main/java/exam/demo/server/pojo/model/Subject.java
@@ -16,19 +16,19 @@ public class Subject extends Model<Subject> {
     /**
      * 雪花算法生成Id
      */
-    private Long id;
+    private Integer id;
     /**
      * 机构id
      */
-    public Long orgId;
+    public Integer orgId;
     /**
      * 机构下公司id
      */
-    public Long companyId;
+    public Integer companyId;
     /**
      * 通过id到数据字典中查询创建者
      */
-    public Long createdBy;
+    public Integer createdBy;
     /**
      * 创建日期
      */
@@ -36,7 +36,7 @@ public class Subject extends Model<Subject> {
     /**
      * 通过id到数据字典中查询修改者
      */
-    public Long updatedBy;
+    public Integer updatedBy;
     /**
      * 修改日期
      */
@@ -48,12 +48,12 @@ public class Subject extends Model<Subject> {
     /**
      * 题目类型id
      */
-    private Long subjectTypeId;
+    private Integer subjectTypeId;
 
     /**
      * 题目类别id
      */
-    private Long categoryId;
+    private Integer categoryId;
 
     /**
      * 题目
@@ -63,7 +63,7 @@ public class Subject extends Model<Subject> {
     /**
      * 难度
      */
-    private Long difficulty;
+    private Integer difficulty;
 
     /**
      * 状态位
@@ -107,7 +107,7 @@ public class Subject extends Model<Subject> {
     private String difficultyName;
 
     @TableField(exist = false)
-    private Long judgeId;
+    private Integer judgeId;
 
     @TableField(exist = false)
     private Long oldVersion;
